Added is_method option
I added an option 'is_method' of boolean type to the Form\Element\Proxy.
If is_method is false, everything works pretty the same as the current
solution. But if is_method is true, the propery acts like a getter
method, so you can use such getters that are not for fields in the
datebase.