Added custom repository method feature
The code is updated to support a custom repository method. The repository method can provided in the options array as follows:

    'repository_method' => 'string'

In the constructor the provided value will be validated and if the variable is not a string or a non existing method an `InvalidArgumentException` will be thrown.

If no custom method is provided the class will fallback to the default `findOneBy` method.