Drop nightly builds as gives zero value feedback
For what I understand reading the history, nightly was something used to
represent the next PHP version. This was a good intention, but at the
end Travis builds as nightly has very little in common with whatever
lands as a final release. So building against nightly is only a waste of
electricity, and no one reads its log because it gives zero value:
nightly errors may disappears once new PHP version is published.